Reconnect Graph in between Lie-steps
In Lie-averaging, relative transformation with an angular error >5 degrees are removed for the second round of averaging. Sometimes too much connections are removed and the graph is not connected anymore. Now we check whether the tree is still connection. If this is not the case, connections with the lowest error are sequentially added until the tree is connection.