Update particles correctly in bootstrapping
In one2all and one2all_symmetric, the particles that are registrered to the template are not updates in 'Particles'. In this way, the orientation of the particles does not improve over the iterations, because every iteration the particles start with the orientation they had in the beginning. This is corrected here.